fixed some bugs

+++ error.c
@@ -6,7 +6,7 @@
 /*   By: naharagu <naharagu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/26 15:06:57 by naharagu          #+#    #+#             */
-/*   Updated: 2022/12/27 12:04:01 by naharagu         ###   ########.fr       */
+/*   Updated: 2022/12/27 15:25:59 by naharagu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	put_error_and_exit(int error_num)
 	exit(EXIT_FAILURE);
 }
 
-void	free_all(t_info *info, size_t i)
+void	free_all(t_info *info, int i)
 {
 	// while (i > 0)
 	// 	free(map[--i]);
